+++ Panther/include/Token.h
@@ -60,8 +60,10 @@ namespace pcit::panther{
 				KeywordVar,
 				KeywordFunc,
 
+				KeywordNull,
 				KeywordUninit,
-				KeywordAddr,
+				
+				// KeywordAddr,
 				KeywordCopy,
 				KeywordMove,
 				KeywordAs,
@@ -70,8 +72,28 @@ namespace pcit::panther{
 				///////////////////////////////////
 				// operators
 
-				Assign, // =
 				RightArrow, // ->
+
+				// assignment
+				Assign, // =
+				AssignAdd, // +=
+				AssignAddWrap, // +@=
+				AssignAddSat, // +|=
+				AssignSub, // -=
+				AssignSubWrap, // -@=
+				AssignSubSat, // -|=
+				AssignMul, // *=
+				AssignMulWrap, // *@=
+				AssignMulSat, // *|=
+				AssignDiv, // /=
+				AssignMod, // %=
+				AssignShiftLeft, // <<=
+				AssignShiftLeftSat, // <<|=
+				AssignShiftRight, // >>=
+				AssignBitwiseAnd, // &=
+				AssignBitwiseOr, // |=
+				AssignBitwiseXOr, // ^=
+
 
 				// arithmetic
 				Plus, // +
@@ -101,13 +123,17 @@ namespace pcit::panther{
 
 				// bitwise
 				ShiftLeft, // <<
-				SatShiftLeft, // <<|
+				ShiftLeftSat, // <<|
 				ShiftRight, // >>
 				BitwiseAnd, // &
 				BitwiseOr, // |
 				BitwiseXOr, // ^
 				BitwiseNot, // ~
 
+				// accessors
+				Accessor, // .
+				Dereference, // .*
+				Unwrap, // .?
 				
 				///////////////////////////////////
 				// punctuation
@@ -121,6 +147,7 @@ namespace pcit::panther{
 				Comma, // ,
 				SemiColon, // ;
 				Colon, // :
+				QuestionMark, // ?
 			};
 
 			struct Location{
@@ -191,11 +218,32 @@ namespace pcit::panther{
 
 			// TODO: hash-map optimization?
 			EVO_NODISCARD static constexpr auto lookupKind(std::string_view op_str) noexcept -> Kind {
+				// length 4
+				if(op_str == "<<|="){ return Kind::AssignShiftLeftSat; }
+
 				// length 3
-				if(op_str == "<<|"){ return Kind::SatShiftLeft; }
+				if(op_str == "<<|"){ return Kind::ShiftLeftSat; }
+				if(op_str == "+@="){ return Kind::AssignAddWrap; }
+				if(op_str == "+|="){ return Kind::AssignAddSat; }
+				if(op_str == "-@="){ return Kind::AssignSubWrap; }
+				if(op_str == "-|="){ return Kind::AssignSubSat; }
+				if(op_str == "*@="){ return Kind::AssignMulWrap; }
+				if(op_str == "*|="){ return Kind::AssignMulSat; }
+				if(op_str == "<<="){ return Kind::AssignShiftLeft; }
+				if(op_str == ">>="){ return Kind::AssignShiftRight; }
+
 
 				// length 2
 				if(op_str == "->"){ return Kind::RightArrow; }
+
+				if(op_str == "+="){ return Kind::AssignAdd; }
+				if(op_str == "-="){ return Kind::AssignSub; }
+				if(op_str == "*="){ return Kind::AssignMul; }
+				if(op_str == "/="){ return Kind::AssignDiv; }
+				if(op_str == "%="){ return Kind::AssignMod; }
+				if(op_str == "&="){ return Kind::AssignBitwiseAnd; }
+				if(op_str == "|="){ return Kind::AssignBitwiseOr; }
+				if(op_str == "^="){ return Kind::AssignBitwiseXOr; }
 
 				if(op_str == "+@"){ return Kind::AddWrap; }
 				if(op_str == "+|"){ return Kind::AddSat; }
@@ -214,6 +262,9 @@ namespace pcit::panther{
 
 				if(op_str == "<<"){ return Kind::ShiftLeft; }
 				if(op_str == ">>"){ return Kind::ShiftRight; }
+
+				if(op_str == ".*"){ return Kind::Dereference; }
+				if(op_str == ".?"){ return Kind::Unwrap; }
 
 
 				// length 1
@@ -235,6 +286,8 @@ namespace pcit::panther{
 				if(op_str == "^"){ return Kind::BitwiseXOr; }
 				if(op_str == "~"){ return Kind::BitwiseNot; }
 
+				if(op_str == "."){ return Kind::Accessor; }
+
 				if(op_str == "("){ return Kind::OpenParen; }
 				if(op_str == ")"){ return Kind::CloseParen; }
 				if(op_str == "["){ return Kind::OpenBracket; }
@@ -244,6 +297,7 @@ namespace pcit::panther{
 				if(op_str == ","){ return Kind::Comma; }
 				if(op_str == ";"){ return Kind::SemiColon; }
 				if(op_str == ":"){ return Kind::Colon; }
+				if(op_str == "?"){ return Kind::QuestionMark; }
 
 				evo::debugFatalBreak("Unknown or unsupported kind ({})", op_str);
 			};
@@ -251,98 +305,127 @@ namespace pcit::panther{
 
 			EVO_NODISCARD static auto printKind(Kind kind) noexcept -> std::string_view {
 				switch(kind){
-					break; case Kind::None:             return "None";
+					break; case Kind::None:               return "None";
 
-					break; case Kind::Ident:            return "Ident";
-					break; case Kind::Intrinsic:        return "Intrinsic";
-					break; case Kind::Attribute:        return "Attribute";
+					break; case Kind::Ident:              return "Ident";
+					break; case Kind::Intrinsic:          return "Intrinsic";
+					break; case Kind::Attribute:          return "Attribute";
 
 
 					///////////////////////////////////
 					// literals
 
-					break; case Kind::LiteralInt:       return "LiteralInt";
-					break; case Kind::LiteralFloat:     return "LiteralFloat";
-					break; case Kind::LiteralBool:      return "LiteralBool";
-					break; case Kind::LiteralString:    return "LiteralString";
-					break; case Kind::LiteralChar:      return "LiteralChar";
+					break; case Kind::LiteralInt:         return "LiteralInt";
+					break; case Kind::LiteralFloat:       return "LiteralFloat";
+					break; case Kind::LiteralBool:        return "LiteralBool";
+					break; case Kind::LiteralString:      return "LiteralString";
+					break; case Kind::LiteralChar:        return "LiteralChar";
+
 
 
 					///////////////////////////////////
 					// types
 
-					break; case Kind::TypeVoid:         return "Void";
-					break; case Kind::TypeType:         return "Type";
-					break; case Kind::TypeInt:          return "Int";
-					break; case Kind::TypeBool:         return "Bool";
+					break; case Kind::TypeVoid:           return "Void";
+					break; case Kind::TypeType:           return "Type";
+					break; case Kind::TypeInt:            return "Int";
+					break; case Kind::TypeBool:           return "Bool";
 
 
 					///////////////////////////////////
 					// keywords
 
-					break; case Kind::KeywordVar:       return "var";
-					break; case Kind::KeywordFunc:      return "func";
+					break; case Kind::KeywordVar:         return "var";
+					break; case Kind::KeywordFunc:        return "func";
 
-					break; case Kind::KeywordUninit:    return "uninit";
-					break; case Kind::KeywordAddr:      return "addr";
-					break; case Kind::KeywordCopy:      return "copy";
-					break; case Kind::KeywordMove:      return "move";
-					break; case Kind::KeywordAs:        return "as";
+					break; case Kind::KeywordNull:        return "null";
+					break; case Kind::KeywordUninit:      return "uninit";
+
+					// break; case Kind::KeywordAddr:     return "addr";
+					break; case Kind::KeywordCopy:        return "copy";
+					break; case Kind::KeywordMove:        return "move";
+					break; case Kind::KeywordAs:          return "as";
 
 
 					///////////////////////////////////
 					// operators
 
-					break; case Kind::Assign:           return "=";
-					break; case Kind::RightArrow:       return "->";
+					break; case Kind::RightArrow:         return "->";
+
+					// assignment
+					break; case Kind::Assign:             return "=";
+					break; case Kind::AssignAdd:          return "+=";
+					break; case Kind::AssignAddWrap:      return "+@=";
+					break; case Kind::AssignAddSat:       return "+|=";
+					break; case Kind::AssignSub:          return "-=";
+					break; case Kind::AssignSubWrap:      return "-@=";
+					break; case Kind::AssignSubSat:       return "-|=";
+					break; case Kind::AssignMul:          return "*=";
+					break; case Kind::AssignMulWrap:      return "*@=";
+					break; case Kind::AssignMulSat:       return "*|=";
+					break; case Kind::AssignDiv:          return "/=";
+					break; case Kind::AssignMod:          return "%=";
+					break; case Kind::AssignShiftLeft:    return "<<=";
+					break; case Kind::AssignShiftLeftSat: return "<<|=";
+					break; case Kind::AssignShiftRight:   return ">>=";
+					break; case Kind::AssignBitwiseAnd:   return "&=";
+					break; case Kind::AssignBitwiseOr:    return "|=";
+					break; case Kind::AssignBitwiseXOr:   return "^=";
 
 					// arithmetic
-					break; case Kind::Plus:             return "+";
-					break; case Kind::AddWrap:          return "+@";
-					break; case Kind::AddSat:           return "+|";
-					break; case Kind::Minus:            return "-";
-					break; case Kind::SubWrap:          return "-@";
-					break; case Kind::SubSat:           return "-|";
-					break; case Kind::Asterisk:         return "*";
-					break; case Kind::MulWrap:          return "*@";
-					break; case Kind::MulSat:           return "*|";
-					break; case Kind::ForwardSlash:     return "/";
-					break; case Kind::Mod:              return "%";
+					break; case Kind::Plus:               return "+";
+					break; case Kind::AddWrap:            return "+@";
+					break; case Kind::AddSat:             return "+|";
+					break; case Kind::Minus:              return "-";
+					break; case Kind::SubWrap:            return "-@";
+					break; case Kind::SubSat:             return "-|";
+					break; case Kind::Asterisk:           return "*";
+					break; case Kind::MulWrap:            return "*@";
+					break; case Kind::MulSat:             return "*|";
+					break; case Kind::ForwardSlash:       return "/";
+					break; case Kind::Mod:                return "%";
 
 					// logical
-					break; case Kind::Equal:            return "==";
-					break; case Kind::NotEqual:         return "!=";
-					break; case Kind::LessThan:         return "<";
-					break; case Kind::LessThanEqual:    return "<=";
-					break; case Kind::GreaterThan:      return ">";
-					break; case Kind::GreaterThanEqual: return ">=";
+					break; case Kind::Equal:              return "==";
+					break; case Kind::NotEqual:           return "!=";
+					break; case Kind::LessThan:           return "<";
+					break; case Kind::LessThanEqual:      return "<=";
+					break; case Kind::GreaterThan:        return ">";
+					break; case Kind::GreaterThanEqual:   return ">=";
 
 					// logical
-					break; case Kind::Not:              return "!";
-					break; case Kind::And:              return "&&";
-					break; case Kind::Or:               return "||";
+					break; case Kind::Not:                return "!";
+					break; case Kind::And:                return "&&";
+					break; case Kind::Or:                 return "||";
 					
 					// bitwise
-					break; case Kind::ShiftLeft:        return "<<";
-					break; case Kind::SatShiftLeft:     return "<<|";
-					break; case Kind::ShiftRight:       return ">>";
-					break; case Kind::BitwiseAnd:       return "&";
-					break; case Kind::BitwiseOr:        return "|";
-					break; case Kind::BitwiseXOr:       return "^";
-					break; case Kind::BitwiseNot:       return "~";
+					break; case Kind::ShiftLeft:          return "<<";
+					break; case Kind::ShiftLeftSat:       return "<<|";
+					break; case Kind::ShiftRight:         return ">>";
+					break; case Kind::BitwiseAnd:         return "&";
+					break; case Kind::BitwiseOr:          return "|";
+					break; case Kind::BitwiseXOr:         return "^";
+					break; case Kind::BitwiseNot:         return "~";
+
+					// Accessors
+					break; case Kind::Accessor:           return ".";
+					break; case Kind::Dereference:        return ".*";
+					break; case Kind::Unwrap:             return ".?";
+
 
 					///////////////////////////////////
 					// punctuation
 
-					break; case Kind::OpenParen:        return "(";
-					break; case Kind::CloseParen:       return ")";
-					break; case Kind::OpenBracket:      return "[";
-					break; case Kind::CloseBracket:     return "]";
-					break; case Kind::OpenBrace:        return "{";
-					break; case Kind::CloseBrace:       return "}";
-					break; case Kind::Comma:            return ",";
-					break; case Kind::SemiColon:        return ";";
-					break; case Kind::Colon:            return ":";
+					break; case Kind::OpenParen:          return "(";
+					break; case Kind::CloseParen:         return ")";
+					break; case Kind::OpenBracket:        return "[";
+					break; case Kind::CloseBracket:       return "]";
+					break; case Kind::OpenBrace:          return "{";
+					break; case Kind::CloseBrace:         return "}";
+					break; case Kind::Comma:              return ",";
+					break; case Kind::SemiColon:          return ";";
+					break; case Kind::Colon:              return ":";
+					break; case Kind::QuestionMark:       return "?";
 				};
 
 				evo::debugFatalBreak("Unknown or unsupported token kind ({})", evo::to_underlying(kind));